requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Async Task Queue
celery==5.3.4
redis==5.0.1

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests (for webhooks)
requests==2.31.0

# Production Server
gunicorn==21.2.0
